[SUPPLICANT] [5/8] Fix PTK rekeying to generate a new ANonce
Supplicant security advisories: https://w1.fi/security/2017-1/
WPA packet number reuse with replayed messages and key reinstallation

[PATCH 5/8] Fix PTK rekeying to generate a new ANonce

The Authenticator state machine path for PTK rekeying ended up bypassing
the AUTHENTICATION2 state where a new ANonce is generated when going
directly to the PTKSTART state since there is no need to try to
determine the PMK again in such a case. This is far from ideal since the
new PTK would depend on a new nonce only from the supplicant.

Fix this by generating a new ANonce when moving to the PTKSTART state
for the purpose of starting new 4-way handshake to rekey PTK.

SCSC-Bug-Id: S17SWF-539

Change-Id: I8ab7860a0a53f88c54cc12289404c6871c84d8ff
Signed-off-by: Anoop Nair <anoop.nair@samsung.com>